backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
groq==0.9.0
httpx<0.28.0
sentence-transformers==2.7.0
pdfplumber==0.11.0
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
cachetools==5.3.3
pydantic==2.7.1
openpyxl==3.1.2
torch
transformers
